+++ src/modules/dateTime/directives/index.ts
@@ -1,6 +1,7 @@
 export * from './dateTime/dateTime.directive';
 export * from './dateTimeBase';
 export * from './dateTimeControlValueAccessor/dateTimeControlValueAccessor.directive';
+export * from './dateTimeFormControl/dateTimeFormControl.directive';
 export * from './dateTimeInput/dateTimeInput.directive';
 export * from './dateTimeMaxValidator/dateTimeMaxValidator.directive';
 export * from './dateTimeMinValidator/dateTimeMinValidator.directive';